consolidate save and update

+++ src/main/java/com/blizzard/todoapi/services/TodoSvc.java
@@ -2,12 +2,11 @@ package com.blizzard.todoapi.services;
 
 import com.blizzard.todoapi.models.Todo;
 import com.blizzard.todoapi.repository.ITodoRepository;
+import graphql.com.google.common.base.Strings;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
-import java.util.Optional;
 
 @Service
 public class TodoSvc {
@@ -18,29 +17,16 @@ public class TodoSvc {
         return todoRepo.findAll();
     }
 
-    public List<Todo> saveTodo(Todo todo) {
+    public List<Todo> saveAndUpdateTodo(Todo todo) {
         todoRepo.save(todo);
         return todoRepo.findAll();
     }
 
-    public List<Todo> updateTodo(String id, Todo todo) {
-        try {
-            Todo tmp = todoRepo.findById(id).get();
-
-            if (tmp != null) {
-                tmp.setTask(todo.getTask());
-                tmp.setDone(todo.isDone());
-
-                todoRepo.save(tmp);
-            }
-        } catch (NoSuchElementException e) {
-            System.out.println("No such element");
+    public List<Todo> deleteTodo(String id) {
+        if (Strings.isNullOrEmpty(id)) {
+            throw new NullPointerException("Must provide a Todo ID");
         }
 
-        return todoRepo.findAll();
-    }
-
-    public List<Todo> deleteTodo(String id) {
         todoRepo.deleteById(id);
         return todoRepo.findAll();
     }
